feat: abort timeout toast
