+++ apps/web/src/lib/utils.ts
@@ -19,6 +19,21 @@ export function formatDateTime(date: Date) {
   return format(date, "LLL dd, y HH:mm:ss");
 }
 
+export function formatDuration(ms: number) {
+  if (ms < 0) ms = -ms;
+  const time = {
+    day: Math.floor(ms / 86400000),
+    hour: Math.floor(ms / 3600000) % 24,
+    min: Math.floor(ms / 60000) % 60,
+    sec: Math.floor(ms / 1000) % 60,
+    ms: Math.floor(ms) % 1000,
+  };
+  return Object.entries(time)
+    .filter((val) => val[1] !== 0)
+    .map(([key, val]) => `${val} ${key}${val !== 1 && key !== "ms" ? "s" : ""}`)
+    .join(", ");
+}
+
 export function notEmpty<TValue>(
   value: TValue | null | undefined,
 ): value is TValue {
